Extract 'cumulative_percentage()' from algorithm V0
